# kratos/api/system/v1/captcha_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v3.19.4
// source: system/v1/captcha.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationCaptchaServiceGenerateImageCaptcha = "/api.system.v1.CaptchaService/GenerateImageCaptcha"
const OperationCaptchaServiceGetEnabledTypes = "/api.system.v1.CaptchaService/GetEnabledTypes"
const OperationCaptchaServiceSendEmailCaptcha = "/api.system.v1.CaptchaService/SendEmailCaptcha"
const OperationCaptchaServiceSendSMSCaptcha = "/api.system.v1.CaptchaService/SendSMSCaptcha"

type CaptchaServiceHTTPServer interface {
	GenerateImageCaptcha(context.Context, *GenerateImageCaptchaRequest) (*CaptchaReply, error)
	GetEnabledTypes(context.Context, *GetEnabledTypesRequest) (*GetEnabledTypesReply, error)
	SendEmailCaptcha(context.Context, *SendEmailCaptchaRequest) (*CaptchaReply, error)
	SendSMSCaptcha(context.Context, *SendSMSCaptchaRequest) (*CaptchaReply, error)
}

func RegisterCaptchaServiceHTTPServer(s *http.Server, srv CaptchaServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/captcha/image", _CaptchaService_GenerateImageCaptcha0_HTTP_Handler(srv))
	r.POST("/captcha/sms", _CaptchaService_SendSMSCaptcha0_HTTP_Handler(srv))
	r.POST("/captcha/email", _CaptchaService_SendEmailCaptcha0_HTTP_Handler(srv))
	r.GET("/captcha/enabled-types", _CaptchaService_GetEnabledTypes0_HTTP_Handler(srv))
}

func _CaptchaService_GenerateImageCaptcha0_HTTP_Handler(srv CaptchaServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GenerateImageCaptchaRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationCaptchaServiceGenerateImageCaptcha)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GenerateImageCaptcha(ctx, req.(*GenerateImageCaptchaRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CaptchaReply)
		return ctx.Result(200, reply)
	}
}

func _CaptchaService_SendSMSCaptcha0_HTTP_Handler(srv CaptchaServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SendSMSCaptchaRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationCaptchaServiceSendSMSCaptcha)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SendSMSCaptcha(ctx, req.(*SendSMSCaptchaRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CaptchaReply)
		return ctx.Result(200, reply)
	}
}

func _CaptchaService_SendEmailCaptcha0_HTTP_Handler(srv CaptchaServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SendEmailCaptchaRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationCaptchaServiceSendEmailCaptcha)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SendEmailCaptcha(ctx, req.(*SendEmailCaptchaRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CaptchaReply)
		return ctx.Result(200, reply)
	}
}

func _CaptchaService_GetEnabledTypes0_HTTP_Handler(srv CaptchaServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetEnabledTypesRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationCaptchaServiceGetEnabledTypes)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetEnabledTypes(ctx, req.(*GetEnabledTypesRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetEnabledTypesReply)
		return ctx.Result(200, reply)
	}
}

type CaptchaServiceHTTPClient interface {
	GenerateImageCaptcha(ctx context.Context, req *GenerateImageCaptchaRequest, opts ...http.CallOption) (rsp *CaptchaReply, err error)
	GetEnabledTypes(ctx context.Context, req *GetEnabledTypesRequest, opts ...http.CallOption) (rsp *GetEnabledTypesReply, err error)
	SendEmailCaptcha(ctx context.Context, req *SendEmailCaptchaRequest, opts ...http.CallOption) (rsp *CaptchaReply, err error)
	SendSMSCaptcha(ctx context.Context, req *SendSMSCaptchaRequest, opts ...http.CallOption) (rsp *CaptchaReply, err error)
}

type CaptchaServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewCaptchaServiceHTTPClient(client *http.Client) CaptchaServiceHTTPClient {
	return &CaptchaServiceHTTPClientImpl{client}
}

func (c *CaptchaServiceHTTPClientImpl) GenerateImageCaptcha(ctx context.Context, in *GenerateImageCaptchaRequest, opts ...http.CallOption) (*CaptchaReply, error) {
	var out CaptchaReply
	pattern := "/captcha/image"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationCaptchaServiceGenerateImageCaptcha))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *CaptchaServiceHTTPClientImpl) GetEnabledTypes(ctx context.Context, in *GetEnabledTypesRequest, opts ...http.CallOption) (*GetEnabledTypesReply, error) {
	var out GetEnabledTypesReply
	pattern := "/captcha/enabled-types"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationCaptchaServiceGetEnabledTypes))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *CaptchaServiceHTTPClientImpl) SendEmailCaptcha(ctx context.Context, in *SendEmailCaptchaRequest, opts ...http.CallOption) (*CaptchaReply, error) {
	var out CaptchaReply
	pattern := "/captcha/email"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationCaptchaServiceSendEmailCaptcha))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *CaptchaServiceHTTPClientImpl) SendSMSCaptcha(ctx context.Context, in *SendSMSCaptchaRequest, opts ...http.CallOption) (*CaptchaReply, error) {
	var out CaptchaReply
	pattern := "/captcha/sms"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationCaptchaServiceSendSMSCaptcha))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}