# native/docs/swagger/swagger.yaml
basePath: /
definitions:
...
  github_com_force-c_nai-tizi_internal_domain_response.Response:
    description: API 统一响应格式
    properties:
      code:
        description: 业务状态码
        example: 200
        type: integer
      data:
        description: 响应数据
        type: object
      msg:
        description: 响应消息
        example: success
        type: string
    type: object
...
  github_com_force-c_nai-tizi_internal_infrastructure_captcha.CaptchaData:
    properties:
      data:
        description: 验证码数据
      expireAt:
        description: 过期时间
        type: string
      id:
        description: 验证码ID
        type: string
      type:
        allOf:
        - $ref: '#/definitions/github_com_force-c_nai-tizi_internal_infrastructure_captcha.CaptchaType'
        description: 验证码类型
    type: object
  github_com_force-c_nai-tizi_internal_infrastructure_captcha.CaptchaType:
    enum:
    - image
    - sms
    - email
    type: string
    x-enum-comments:
      CaptchaTypeEmail: 邮箱验证码
      CaptchaTypeImage: 图形验证码
      CaptchaTypeSMS: 短信验证码
    x-enum-descriptions:
    - 图形验证码
    - 短信验证码
    - 邮箱验证码
    x-enum-varnames:
    - CaptchaTypeImage
    - CaptchaTypeSMS
    - CaptchaTypeEmail
...
  internal_controller.SendEmailCaptchaRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  internal_controller.SendSMSCaptchaRequest:
    properties:
      phone:
        type: string
    required:
    - phone
    type: object
host: localhost:9009
info:
  contact:
    email: support@example.com
    name: 技术支持
    url: https://example.com/support
  description: Nai-Tizi RESTful API 接口文档，支持双 Token 认证机制
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://example.com/terms/
  title: NTZ API 文档
  version: "1.0"
paths:
...
  /captcha/email:
    post:
      consumes:
      - application/json
      parameters:
      - description: 邮箱
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_controller.SendEmailCaptchaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_force-c_nai-tizi_internal_domain_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_force-c_nai-tizi_internal_infrastructure_captcha.CaptchaData'
              type: object
      summary: 发送邮箱验证码
      tags:
      - 验证码
...
  /captcha/sms:
    post:
      consumes:
      - application/json
      parameters:
      - description: 手机号
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_controller.SendSMSCaptchaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_force-c_nai-tizi_internal_domain_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_force-c_nai-tizi_internal_infrastructure_captcha.CaptchaData'
              type: object
      summary: 发送短信验证码
      tags:
      - 验证码
...
securityDefinitions:
  Bearer:
    description: '格式: "Bearer {access_token}"，AccessToken 从登录接口获取'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...